+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.7.39
+
+### Patch Changes
+
+-   Updated dependencies [65bfb244]
+    -   @sap-ux/axios-extension@1.14.2
+
+## 0.7.38
+
+### Patch Changes
+
+-   Updated dependencies [844e79c4]
+    -   @sap-ux/axios-extension@1.14.1
+
+## 0.7.37
+
+### Patch Changes
+
+-   Updated dependencies [31cc53f8]
+    -   @sap-ux/axios-extension@1.14.0
+
 ## 0.7.36
 
 ### Patch Changes